moved `plot/_plot.py` to folder `framework`
